requirements.txt
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.6.0
boto3==1.9.194
botocore==1.12.194
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-debug-toolbar==2.0
django-extensions==2.1.9
django-extra-views==0.11.0
django-haystack==2.8.1
django-modelcluster==4.4
django-oscar==2.0
django-phonenumber-field==2.0.1
django-storages==1.7.1
django-tables2==1.21.2
django-taggit==0.24.0
django-treebeard==4.3
django-widget-tweaks==1.4.5
djangorestframework==3.10.1
docutils==0.14
draftjs-exporter==2.1.6
factory-boy==2.12.0
Faker==2.0.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
jmespath==0.9.4
parso==0.5.1
pexpect==4.7.0
phonenumbers==8.10.15
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
purl==1.5
Pygments==2.4.2
pysolr==3.7.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.5.1
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
Willow==1.1
